+++ NEWS.md
@@ -10,9 +10,10 @@
 
 # OpenRepGrid 0.1.14 (work in progress)
   
+  * summary measures for indexDilemma (PID, IID, PICID)
   * midpoint function (alias for getScaleMidpoint)
   * print method for indexDilemma
-  * correct indexDilemma output bug (#17, thanks to José Antonio González Del Puerto aka @MindCartographer)
+  * fix indexDilemma output bug (#17, thanks to José Antonio González Del Puerto aka @MindCartographer)
   * perturbation of grid ratings by perturbate and grids_perturbate
   * indexPvaff now uses PCA of construct centered raw data in line with biplot (and Gridcor) 
   * allow blanks at end of line after tags in importTxt